rename to winter:test and support core tests
